Info section now centers vertically
jQuery is now used to center the information section vertically in the
banner. The opacity change is now dependent on a ternary operator,
meaning if the value will be below 0 then no change occurs. The 'height'
variable has been renamed to 'offset' as it is not the height but the
difference between the image height and the banner height.